# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
...
  "paths": {
...
    "/city": {
      "post": {
        "tags": ["cities"],
        "summary": "Obtenir une ville via son nom",
        "description": "Permet d'obtenir les informations d'une ville à partir de son nom",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": true  
          }
        ]
      }
    },
...
    "/city/create": {
      "post": {
        "tags": ["cities"],
        "summary": "Créer une ville",
        "description": "Permet de créer une ville",
        "parameters": [
          {
            "name": "department_code",
            "in": "formData",
            "type": "string",
            "required": true,
            "enum": ["01","02","03","04","05","06","07","08","09","10","11","12","13","14","15","16","17","18","19","20","21","22","23","24","25","26"]
          },
          {
            "name": "zip_code",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "gps_lat",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "gps_lng",
            "in": "formData",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/image/create": {
      "post": {
        "tags": ["images"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "image": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/user/create": {
      "post": {
        "tags": ["users"],
        "summary": "Créer un utilisateur",
        "description": "Permet d'ajouter un nouvelle utilisateur",
        "parameters": [
          {
            "name": "firstname",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "Jean"
          },
          {
            "name": "lastname",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "BON"
          },
          {
            "name": "email",
            "in": "formData",
            "type": "string",
            "required": true,
            "uniqueItems": true,
            "default": "test@test.com"

          },
          {
            "name": "password",
            "in": "formData",
            "type": "string",
            "required": true,
            "schema": {
              "pattern": "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[a-zA-Z]).{8,}$"
            },
            "default": "Azerty_1234"

          },
          {
            "name": "status",
            "description": "Le statut de l'utilisateur (actif ou inactif)",
            "in": "formData",
            "type": "boolean",
            "required": true,
            "schema": {
              "type": "boolean",
              "enum": [true, false]
            },
            "default": true

          },
          {
            "name": "role",
            "description": "Le role de l'utilisateur (user ou admin)",
            "in": "formData",
            "type": "string",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["user", "admin"]
            },
            "default": "user"

          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/route": {
      "post": {
        "tags": ["routes"],
        "summary": "Obtenir un trajets via l'id",
        "description": "Permet de trouver un trajets via son id",
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/route/create": {
      "post": {
        "tags": ["routes"],
        "summary": "Créer un trajet",
        "description": "Permet de créer un trajet",
        "parameters": [
          {
            "name": "arrival_city_id",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "departure_city_id",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "departure_time",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "2022-11-11 08:00:00"
          },
          {
            "name": "arrival_time",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "2022-11-11 08:30:00"
          },
          {
            "name": "departure_date",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "2022-11-11"
          },
          {
            "name": "vehicules_id",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "2"
          },
          {
            "name": "available_seats",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "remaining_seats",
            "in": "formData",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/userHasRoute/route": {
      "post": {
        "tags": ["users_has_routes"],
        "summary": "Obtenir les utilisateurs participant à un trajet",
        "description": "Permet de trouver les utilisateurs participant à un trajet",
        "parameters": [
          {
            "name": "route_id",
            "in": "formData",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/userHasRoute/user": {
      "post": {
        "tags": ["users_has_routes"],
        "summary": "Obtenir les trajets d'un utilisateur",
        "description": "Permet d'obtenir les trajets d'un utilisateur",
        "parameters": [
          {
            "name": "user_id",
            "in": "formData",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/userHasRoute/create": {
      "post": {
        "tags": ["users_has_routes"],
        "summary": "Obtenir les trajets via l'id de la ville de depart ou d'arrivée",
        "description": "Permet de trouver tout les trajets ayant comme ville d'arrivée ou de départ ou les deux la ville indiquée",
        "parameters": [
          {
            "name": "user_id",
            "in": "formData",
            "type": "integer",
            "required": true
          },
          {
            "name": "route_id",
            "in": "formData",
            "type": "integer",
            "required": true
          },
          {
            "name": "status_notice",
            "in": "formData",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/model": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "mark": {
                  "example": "any"
                },
                "model": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "tags": ["other"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/forgot": {
      "post": {
        "tags": ["other"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/valide/link": {
      "post": {
        "tags": ["other"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "link": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/forgot/update": {
      "put": {
        "tags": ["other"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password1": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}